Redirect most of the basic module functions to base.
