+++ ompi/mca/coll/base/coll_base_barrier.c
@@ -321,8 +321,8 @@ int ompi_coll_base_barrier_intra_two_procs(struct ompi_communicator_t *comm,
 
 /* copied function (with appropriate renaming) starts here */
 
-static int ompi_coll_base_barrier_intra_basic_linear(struct ompi_communicator_t *comm,
-                                                      mca_coll_base_module_t *module)
+int ompi_coll_base_barrier_intra_basic_linear(struct ompi_communicator_t *comm,
+                                              mca_coll_base_module_t *module)
 {
     int i, err, rank, size;
 
@@ -363,15 +363,14 @@ static int ompi_coll_base_barrier_intra_basic_linear(struct ompi_communicator_t 
         ompi_request_wait_all( size-1, requests+1, MPI_STATUSES_IGNORE );
 
         for (i = 1; i < size; ++i) {
-            err = MCA_PML_CALL(isend(NULL, 0, MPI_BYTE, i,
-                                     MCA_COLL_BASE_TAG_BARRIER,
-                                     MCA_PML_BASE_SEND_STANDARD, comm,
-                                     &(requests[i])));
+            err = MCA_PML_CALL(send(NULL, 0, MPI_BYTE, i,
+                                    MCA_COLL_BASE_TAG_BARRIER,
+                                    MCA_PML_BASE_SEND_STANDARD, comm));
             if (MPI_SUCCESS != err) {
                 return err;
             }
         }
-        ompi_request_wait_all( size-1, requests+1, MPI_STATUSES_IGNORE );
+
         free( requests );
     }
