+++ ompi/mca/coll/base/coll_base_functions.h
@@ -104,10 +104,18 @@ int ompi_coll_base_alltoall_intra_bruck(ALLTOALL_ARGS);
 int ompi_coll_base_alltoall_intra_basic_linear(ALLTOALL_ARGS);
 int ompi_coll_base_alltoall_intra_linear_sync(ALLTOALL_ARGS, int max_requests);
 int ompi_coll_base_alltoall_intra_two_procs(ALLTOALL_ARGS);
+int mca_coll_base_alltoall_intra_basic_inplace(void *rbuf, int rcount,
+                                               struct ompi_datatype_t *rdtype,
+                                               struct ompi_communicator_t *comm,
+                                               mca_coll_base_module_t *module);  /* special version for INPLACE */
 
 /* AlltoAllV */
 int ompi_coll_base_alltoallv_intra_pairwise(ALLTOALLV_ARGS);
 int ompi_coll_base_alltoallv_intra_basic_linear(ALLTOALLV_ARGS);
+int mca_coll_base_alltoallv_intra_basic_inplace(void *rbuf, const int *rcounts, const int *rdisps,
+                                                struct ompi_datatype_t *rdtype,
+                                                struct ompi_communicator_t *comm,
+                                                mca_coll_base_module_t *module);  /* special version for INPLACE */
 
 /* AlltoAllW */
 
@@ -116,8 +124,8 @@ int ompi_coll_base_barrier_intra_doublering(BARRIER_ARGS);
 int ompi_coll_base_barrier_intra_recursivedoubling(BARRIER_ARGS);
 int ompi_coll_base_barrier_intra_bruck(BARRIER_ARGS);
 int ompi_coll_base_barrier_intra_two_procs(BARRIER_ARGS);
-int ompi_coll_base_barrier_intra_linear(BARRIER_ARGS);
 int ompi_coll_base_barrier_intra_tree(BARRIER_ARGS);
+int ompi_coll_base_barrier_intra_basic_linear(BARRIER_ARGS);
 
 /* Bcast */
 int ompi_coll_base_bcast_intra_basic_linear(BCAST_ARGS);
