+++ ompi/mca/coll/basic/coll_basic_bcast.c
@@ -2,7 +2,7 @@
  * Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
  *                         University Research and Technology
  *                         Corporation.  All rights reserved.
- * Copyright (c) 2004-2005 The University of Tennessee and The University
+ * Copyright (c) 2004-2015 The University of Tennessee and The University
  *                         of Tennessee Research Foundation.  All rights
  *                         reserved.
  * Copyright (c) 2004-2005 High Performance Computing Center Stuttgart, 
@@ -28,78 +28,6 @@
 #include "coll_basic.h"
 #include "ompi/mca/pml/pml.h"
 #include "opal/util/bit_ops.h"
-
-
-/*
- *	bcast_lin_intra
- *
- *	Function:	- broadcast using O(N) algorithm
- *	Accepts:	- same arguments as MPI_Bcast()
- *	Returns:	- MPI_SUCCESS or error code
- */
-int
-mca_coll_basic_bcast_lin_intra(void *buff, int count,
-                               struct ompi_datatype_t *datatype, int root,
-                               struct ompi_communicator_t *comm,
-                               mca_coll_base_module_t *module)
-{
-    int i;
-    int size;
-    int rank;
-    int err;
-    ompi_request_t **preq;
-    mca_coll_basic_module_t *basic_module = (mca_coll_basic_module_t*) module;
-    ompi_request_t **reqs = basic_module->mccb_reqs;
-
-    size = ompi_comm_size(comm);
-    rank = ompi_comm_rank(comm);
-
-    /* Non-root receive the data. */
-
-    if (rank != root) {
-        return MCA_PML_CALL(recv(buff, count, datatype, root,
-                                 MCA_COLL_BASE_TAG_BCAST, comm,
-                                 MPI_STATUS_IGNORE));
-    }
-
-    /* Root sends data to all others. */
-
-    for (i = 0, preq = reqs; i < size; ++i) {
-        if (i == rank) {
-            continue;
-        }
-
-        err = MCA_PML_CALL(isend_init(buff, count, datatype, i,
-                                      MCA_COLL_BASE_TAG_BCAST,
-                                      MCA_PML_BASE_SEND_STANDARD,
-                                      comm, preq++));
-        if (MPI_SUCCESS != err) {
-            return err;
-        }
-    }
-    --i;
-
-    /* Start your engines.  This will never return an error. */
-
-    MCA_PML_CALL(start(i, reqs));
-
-    /* Wait for them all.  If there's an error, note that we don't
-     * care what the error was -- just that there *was* an error.  The
-     * PML will finish all requests, even if one or more of them fail.
-     * i.e., by the end of this call, all the requests are free-able.
-     * So free them anyway -- even if there was an error, and return
-     * the error after we free everything. */
-
-    err = ompi_request_wait_all(i, reqs, MPI_STATUSES_IGNORE);
-
-    /* Free the reqs */
-
-    mca_coll_basic_free_reqs(reqs, i);
-
-    /* All done */
-
-    return err;
-}
 
 
 /*
