+++ ompi/mca/coll/basic/coll_basic_gather.c
@@ -2,7 +2,7 @@
  * Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
  *                         University Research and Technology
  *                         Corporation.  All rights reserved.
- * Copyright (c) 2004-2005 The University of Tennessee and The University
+ * Copyright (c) 2004-2015 The University of Tennessee and The University
  *                         of Tennessee Research Foundation.  All rights
  *                         reserved.
  * Copyright (c) 2004-2005 High Performance Computing Center Stuttgart, 
@@ -26,68 +26,6 @@
 #include "ompi/mca/coll/coll.h"
 #include "ompi/mca/coll/base/coll_tags.h"
 #include "ompi/mca/pml/pml.h"
-
-/*
- *	gather_intra
- *
- *	Function:	- basic gather operation
- *	Accepts:	- same arguments as MPI_Gather()
- *	Returns:	- MPI_SUCCESS or error code
- */
-int
-mca_coll_basic_gather_intra(void *sbuf, int scount,
-                            struct ompi_datatype_t *sdtype,
-                            void *rbuf, int rcount,
-                            struct ompi_datatype_t *rdtype,
-                            int root, struct ompi_communicator_t *comm,
-                            mca_coll_base_module_t *module)
-{
-    int i;
-    int err;
-    int rank;
-    int size;
-    char *ptmp;
-    MPI_Aint incr;
-    MPI_Aint extent;
-    MPI_Aint lb;
-
-    size = ompi_comm_size(comm);
-    rank = ompi_comm_rank(comm);
-
-    /* Everyone but root sends data and returns. */
-
-    if (rank != root) {
-        return MCA_PML_CALL(send(sbuf, scount, sdtype, root,
-                                 MCA_COLL_BASE_TAG_GATHER,
-                                 MCA_PML_BASE_SEND_STANDARD, comm));
-    }
-
-    /* I am the root, loop receiving the data. */
-
-    ompi_datatype_get_extent(rdtype, &lb, &extent);
-    incr = extent * rcount;
-    for (i = 0, ptmp = (char *) rbuf; i < size; ++i, ptmp += incr) {
-        if (i == rank) {
-            if (MPI_IN_PLACE != sbuf) {
-                err = ompi_datatype_sndrcv(sbuf, scount, sdtype,
-                                      ptmp, rcount, rdtype);
-            } else {
-                err = MPI_SUCCESS;
-            }
-        } else {
-            err = MCA_PML_CALL(recv(ptmp, rcount, rdtype, i,
-                                    MCA_COLL_BASE_TAG_GATHER,
-                                    comm, MPI_STATUS_IGNORE));
-        }
-        if (MPI_SUCCESS != err) {
-            return err;
-        }
-    }
-
-    /* All done */
-
-    return MPI_SUCCESS;
-}
 
 
 /*
